requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bio==1.7.1
biopython==1.84
biothings-client==0.3.1
blinker==1.8.2
boto3==1.35.10
botocore==1.35.10
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
gprofiler-official==1.0.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
minio==7.2.8
mygene==3.2.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
platformdirs==4.2.2
pooch==1.8.2
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4